(pd) minimal incorporation of ingred/labware setup into PD
