+++ protocol-designer/src/components/App.js
@@ -1,7 +1,8 @@
+// @flow
 import React from 'react'
 import { HashRouter, Route } from 'react-router-dom'
 
-import Home from './Home'
+import ConnectedDeckSetup from '../containers/ConnectedDeckSetup'
 import About from './About'
 import ProtocolEditor from './ProtocolEditor'
 
@@ -15,7 +16,7 @@ export default function App () {
     <HashRouter>
       <div className='container'>
         <Route exact path='/' component={ProtocolEditor} />
-        <Route exact path='/old-ingredient-selector' component={Home} />
+        <Route exact path='/old-ingredient-selector' component={ConnectedDeckSetup} />
         <Route path='/about' component={About} />
       </div>
     </HashRouter>
